backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
python-dotenv>=1.0.0
dspy>=2.5.0
zhipuai>=2.1.0
httpx>=0.28.0
numpy>=2.1.0
shapely>=2.0.0
Pillow>=11.0.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
